Prevent deletion of questions used in question groups. 	Change on 2013/06/27 by johncox <johncox@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=48643429

+++ coursebuilder/modules/dashboard/question_editor.py
@@ -147,6 +147,17 @@ class BaseQuestionRESTHandler(BaseRESTHandler):
             transforms.send_json_response(
                 self, 404, 'Question not found.', {'key': key})
             return
+
+        used_by = QuestionDAO.used_by(question.id)
+        if used_by:
+            group_names = ['"%s"' % x for x in used_by]
+            transforms.send_json_response(
+                self, 403,
+                ('Question in use by question groups:\n%s.\nPlease delete it '
+                 'from those groups and try again.') % ',\n'.join(group_names),
+                {'key': key})
+            return
+
         QuestionDAO.delete(question)
         transforms.send_json_response(self, 200, 'Deleted.')
